Final Edit ???

+++ src/uni/dc/ubsOpti/goataaExt/algorithms/HillClimbingTraceable.java
@@ -106,7 +106,7 @@ public final class HillClimbingTraceable<G, X> extends LocalSearchAlgorithmTrace
 			pnew.v = f.compute(pnew.x, r);
 			// break if the algo is stuck
 			if (Arrays.equals((int[]) pnew.x, (int[]) p.x)) {
-				step = Integer.valueOf(term.getConfiguration(false)) + 1;
+				step = Long.valueOf(term.getConfiguration(false)) + 1;
 				break;
 			}
 			notifyTracer(pnew, p);
